[Items V2] Add example v2 importer
I've thought of a better way we can import items into the game by using the classicdb.ch website to pull all the data in for us, the only thing we need to do is add a description and the url of the item.

Risk should be randomly assigned with there being more lower risk items in the game.